+++ src/main/kotlin/io/openfuture/chain/consensus/service/Services.kt
@@ -1,6 +1,7 @@
 package io.openfuture.chain.consensus.service
 
 import io.openfuture.chain.core.model.entity.Delegate
+import io.openfuture.chain.core.model.entity.block.MainBlock
 
 interface EpochService {
 
@@ -23,5 +24,7 @@ interface EpochService {
     fun getEpochEndTime(): Long
 
     fun getFullTimeSlotDuration(): Long
+
+    fun getEpochByBlock(block: MainBlock): Long
 
 }