+++ src/main/kotlin/io/openfuture/chain/core/service/block/BaseBlockService.kt
@@ -1,5 +1,6 @@
 package io.openfuture.chain.core.service.block
 
+import io.openfuture.chain.core.component.BlockCapacityChecker
 import io.openfuture.chain.core.exception.ValidationException
 import io.openfuture.chain.core.model.entity.block.Block
 import io.openfuture.chain.core.model.entity.block.payload.BlockPayload
@@ -18,11 +19,13 @@ abstract class BaseBlockService<T : Block>(
     protected val repository: BlockRepository<T>,
     protected val blockService: BlockService,
     private val walletService: WalletService,
-    protected val delegateService: DelegateService
+    protected val delegateService: DelegateService,
+    private val  capacityChecker: BlockCapacityChecker
 ) {
 
     protected fun save(block: T): T {
         updateBalanceByReward(block)
+        capacityChecker.incrementCapacity(block.timestamp)
         return repository.save(block)
     }